+++ src/components/RechargeMaterialComponent.js
@@ -20,8 +20,6 @@ class RechargeMaterialComponent {
   bindEventListener() {
     this.$rechargeDrinkButtonContainer.addEventListener('click', this.onRechargeButtonClick);
   }
-  display() {}
-  hidden() {}
   onRechargeButtonClick = e => {
     e.preventDefault();
     if (e.target.id === 'recharge-coffee-beans-button') {
