+++ src/index.js
@@ -1,5 +1,3 @@
 import './styles/index';
 import CoffeeMachineComponent from './components/CoffeeMachineComponent';
-
-localStorage.clear();
-new CoffeeMachineComponent();
+const coffeeMachine = new CoffeeMachineComponent();
